[FIX] l10n_us_payment_nacha: count cents in file control record
Previously the amounts were incorrectly rounded to integer USD
amounts.

It seems that most banks don't actually check if all entries in the
file sum up to what's specified in the file control record. They
instead ignore this part of the file. This is why this bug had gone
unnoticed for so long. We still choose to fix it because it caused
some confusion.

opw-2858885

closes odoo/enterprise#28110

X-original-commit: 53bbb16d60e030f4143931949425fa94d2f7b928
Signed-off-by: jorenvo <jov@odoo.com>