[IMP] account_reports: journal audit fix
When opening the report, we are checking if any journal has been
specifically selected to unfold it.
But as we can't be completely sure of what is in the options given
to the report, as they can be from another report, there can
be some issue in specific cases (for example, journal groups in the
journal options).
Thus, replace the condition with a get to better handle use cases
where the selected key could not be in every line of the journal
options.

closes odoo/enterprise#28068

Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>