[FIX] l10n_fr_fec_import: Montantdevise correctly imported
How to reproduce the bug ?

- install accounting,l10n_fr,l10n_fr_fec and l10n_fr_fec_import
- if necessary create a french company using euro
- switch to the french company
- in settings -> accounting, enable USD and update manually the rate
- in accounting -> accounting -> configuration -> fec import, select and
import the fec file
- go to accounting -> accounting -> journal entries annd search for Fec
journal

What is the bug ?

When you try to import a FEC file into the accouting module, the foreign
currency is well read. This is due to the fact that there is a typo in
the name of the column ("Montantdevise" instead of "MontantDevise").
According to the odoo documentation, the column in the file is well
written.

opw-2853137

closes odoo/enterprise#28105

X-original-commit: 5b46b39b53e1288611b9e90fb127c3fa6560de46
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>