[FIX] timesheet_grid: fix the traceback in mail template
Currently, changing the employee in a mail template throws an error
and it happens because the round function doesn't get the correct
value of the 'timesheet_hours' field.

So in this commit, we fix the issue by checking the value of the
field and if context does not exist then pass zero so changing
the employee in the mail template does not throw an error.

task-2877225

closes odoo/enterprise#28443

X-original-commit: 8b56f275f9a9388214ed69ca16f212573929469a
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>