[FIX] sale_subscription_dashboard: properly display controlpanel content
Problem: the sale subcription dashboard main controlpanel was broken
since a recent owl update. It worked before by chance, because Owl would
render too much the components, but since then, the control panel
content is not displayed. This is because it is rendered after the
controlpanel is rendered itself, so the content is not taken into
account.

With this commit, we force an initial render of the search view and the
buttons, in the `start` method, so it will be properly used by the
controlpanel when it is rendered initially.

This commit also fix an issue: it was not possible to click on a row
in the salesmen dashboard to display the corresponding sale order.

This commit also add some cleanup.

Taskid: 2855699
X-original-commit: 28cb827e80ef8152d396054c6d3189b83ef16630
Part-of: odoo/enterprise#28509