[FIX] account_bank_statement_import_camt: correct amount with charges
This is a small fixup of recent commit [1].

It seems some files do include the charges in the Amt element at
NtryDtls level, while others don't. The mentioned commit aims to always
retrieve the amount at Ntry level in case of charges (as it always
includes them) but this is creating side effects in case of batch
transactions. We should partially revert it but still manage the case
where there is only one transaction and charges are not included.

[1]:https://github.com/odoo/enterprise/commit/b10c498f57b36b1f194f0064539d5fb39bdc27c2

closes odoo/enterprise#28276

X-original-commit: 9a67fa01ecd66d76537d022b5b3d5a456e6e5756
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Alex Thuyls (alt) <alt@odoo.com>