[FIX] documents_spreadsheet: fix gauge chart lib max point at 0
In a spreadsheet: add a gauge chart with the min range being -1 and
the max 0.
=> the circle spans 90° instead of 180°

The bug comes from gauge lib when the last element in `data` is 0 (i.e. when the range maximum is 0).
The value wrongly fallbacks to 1 because 0 is falsy [1]

See the proposed fix: https://github.com/haiiaaa/chartjs-gauge/pull/33

[1] https://github.com/haiiaaa/chartjs-gauge/blob/2ea50541d754d710cb30c2502fa690ac5dc27afd/src/controllers/controller.gauge.js#L52

Part-of: odoo/enterprise#28141