[FIX] account_bank_statement_import_csv: fix problem with import action
A recent update in owl (see [1]) broke a few screens that were working
by chance, so the fix needs to be done manually at each place. One of
them was the import client action. A fix was done in odoo community (see
[2]) in saas 15.3. However, in the meantime, wan fixed it in master (see
[3]). Funnily enough, wan fix worked because the fix in community was
not there yet, but broke after the saas 15.3 fix was forward ported to
master.

This commit should hopefully fix the situation by removing the specific
code that caused the issue (by rendering twice the buttons).

[1] https://github.com/odoo/odoo/commit/7fc0cc7c3f1c770dd80e8c3f3316c249d0708fe4
[2] https://github.com/odoo/odoo/commit/87ed548bc9508ff2e74e5c6cbab8a393646957d5
[3] https://github.com/odoo/enterprise/pull/25123/commits/ed5cb36554931cd789e527fe5563fb398d47af96

closes odoo/enterprise#28157

X-original-commit: caf2f2062d65edb4f30d3f6348b055c3055bbc55
Signed-off-by: William André (wan) <wan@odoo.com>