[REF] documents_spreadsheet: extract spreadsheet code into new module
This task is a first step towards dashboards with spreadsheet.

As the dashboards should be accessible without Documents, we have to
extract the minimal assets required to display a spreadsheet in dashboard
mode.

In dashboard mode, spreadsheet will only readonly (no edition). The
edition comes with Documents app.

Thus, a new module is introduced: `spreadsheet`. It contains only static
files that will be lazy loaded when the first time a spreadsheet is
displayed. All "edition only" code (plugins + ui) are left in the
`documents_spreadsheet` module

Tests for code in the `spreadsheet` module are also moved to this module.
They were all mixed up (it was a big mess) and we needed to reorganize
all tests

Part of task-id 2799682

closes odoo/enterprise#28355

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>
Co-authored-by: Lucas Lefèvre <lul@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>