[IMP] web_gantt, _*: improve the task gantt's pill title
_* = planning, project_enterprise, services_web_gantt

When the task plan is less than 24 hours, the start hours, end hours, and
allocated hours and the name of the task will appear in the gantt view in
the pill title.

Example- '7:00 AM - 11:00 AM (4h) - DST Task 1'

Currently this features is only in planning.

Create a new services_web_gantt module and it's used in planning and
project_enterpise modules.

task-2672783

closes odoo/enterprise#28372

Related: odoo/documentation#2649
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>