[FIX] quality_mrp: QC for manufactured product erased
When having a Quality Check of type quantity/move_line on a manufactured product, a
QC was created along a move line when the user was inserting the lot of
the manufactured product. After marking the MO as done, the line was
deleted and a new one was created along a new QC, losing the output of
the former QC.
This was fixed by removing the mechanism that created a move line (and
therefore a QC of lot type) when the user was setting the lot of the
manufactured product. Instead, in the case of the manufactured product
we create the QC of lot type the same time as the QC of product type.
The lot number is tracked by linking it to the lot number of the MO
instead of the corresponding move line.

Part-of: odoo/enterprise#27602