[FIX] l10n_de_reports: wrong total value in multicurrency invoice
Steps to reproduce:
- Install accounting with german localization
- Activate multi-currency in settings
- Create an invoice with a currency different than EUR
- Add multiple lines of products on the invoice
- Go to reporting>general ledger and download datev ZIP

Current behavior:
The value in the non EUR currency only corresponds to the first line
of the invoice.

Expected behavior:
The value in the non EUR currency corresponds to the total of the
invoice.

Explanation:
In the fix commit that added the non EUR currency to the datev report
(134978b16fcd62c8da72767830b3323846f86521), a mistake was made and the
code was misunderstood so that only the first line price_total of the
invoice was added to the report. We also add a check on the number of
currencies used by move so that the fields 3,4,5 are only filled when
there is only one currency for every lines and that currency is different
than the company currency.

opw-2842120

closes odoo/enterprise#28487

X-original-commit: 45a8f48180bd824088473043ec003d2867938ffa
Signed-off-by: Cedric Snauwaert <csn@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>