[FIX] documents_spreadsheet: correctly handle fetching errors
There was two issues when handling fetching errors:

1. Call `get` (or `batch.get`) then `fetch` before the rpc is resolved
=> the fetch throws "Data is loading" instead of returning a (pending) promise.
To fix the issue, we completely split sync and async cache. This may
not be optimal but it's simple.

2. Correctly cache one by one retries when a batch fails
The `batch` in the `then` closure was always the batch containing
the last request because it was not correctly scoped.

This commit also adds unit tests

closes odoo/enterprise#28211

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>