[IMP] helpdesk: unarchive tickets when unarchiving stages
Currently, all of the tickets contained in a stage are archived
when the stage is archived. However, the tickets remained archived
when the stage is unarchived.

So in this commit, If the user unarchives a stage, ask the user to
unarchive the tickets it contains in the process.

task-2835624

closes odoo/enterprise#28301

Related: odoo/odoo#94008
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>