[FIX] stock_barcode: no longer increment lots with scanned quantities
Before this commit, the quantity was increased each time a lot number is
scanned. This can lead to inconsistencies if a tracked product line already has
a quantity set but no tracking number (scanning the lot will incorrectly
increment the quantity). As from this commit, the quantity will no longer
change when assigning a tracking number to an existing line which already has a
set quantity.

opw-2870675

closes odoo/enterprise#36789

X-original-commit: fad3fbd7991fa6ba9e4267c5013cc5a948b00e9b
Signed-off-by: Steve Van Essche <svs@odoo.com>