[FIX] stock_barcode: no longer increment lots with scanned quantities
Before this commit, the quantity was increased each time a lot number is
scanned. This can lead to inconsistencies if a tracked product line already has
a quantity set but no tracking number (scanning the lot will incorrectly
increment the quantity). As from this commit, the quantity will no longer
change when assigning a tracking number to an existing line which already has a
set quantity.

opw-2870675

closes odoo/enterprise#34980

X-original-commit: 3c3149513176ff0ff225aa0c67836f122d0abce1
Signed-off-by: Steve Van Essche <svs@odoo.com>