[FIX] sign : false placeholder when no developer mode
Steps to reproduce:
- Install sign
- Put no developer mode on
- Set a signature block on a document
- Change if from customer to employee

Current behavior:
The block shows false

Expected behavior:
The block does not change

Explanation:
The bug was created after the fix commit 860ace5013d7f5cde1018f5eb1ec4fce1f6b3537
the fix did not take into consideration that the o_sign_name of the block form
only appeared when the use is in developer mode (t-if="widget.debug" in xml). So the
name was always false when pulled in the form hide function. To fix that we only pull
the name from o_sign_name if we are not in developer mode.

opw-2877149

Part-of: odoo/enterprise#28303