[IMP] sale_subscription: update the way automatic invoices are handled
Before this commit, when an automatic payment with token failed on the
subscription, mails were sent to the customer. Without action, the cron would
try to do the payment every days.
If it failed everyday, a mail was sent the day the payment failed, 3 days,
after one week and after 2 weeks.
The default behavior was to automatically close the contract 15 days after the
payment failed.

As automatic payment can often fails one day but succeed the day after, we
decided to only send a mail after two failures and adapt the remaining mail
schedule.

Moreover, if a payment token is set on the subscription, even if the payment_mode
is validate_send (post invoice and send it), we try to charge it automatically.
If no token is present, no further action will be needed as the customer is
expected to pay.

Taskid: 285569
X-original-commit: 7644defa2ef8af3859373e8ff292685ddfa26a27
Part-of: odoo/enterprise#28725