[FIX] l10n_be_reports: 281_50 testing form seq
The aim of this commit is to allow the generation of a form 325 test xml
without burning real sequence number.

Without this, any xml generation for a test will require a technical
intervention to reset the sequence in order to create the valid xml
later on.

Before this commit:
Creating an xml for a test 325 form will burn a real sequence number.

After this commit:
Creating an xml for a test will get a correct sequence number without
burning any real sequence number.

closes odoo/enterprise#33226

Task: 2820482
Pr-enterprise: https://github.com/odoo/enterprise/pull/28326
X-original-commit: 74a1aa95de8a002f22b68ef83db86ad722cbeecd
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>