[FIX] l10n_be_reports: 281_50 testing form seq
The aim of this commit is to allow the generation of a form 325 test xml
without burning real sequence number.

Without this, any xml generation for a test will require a technical
intervention to reset the sequence in order to create the valid xml
later on.

Before this commit:
Creating an xml for a test 325 form will burn a real sequence number.

After this commit:
Creating an xml for a test will get a correct sequence number without
burning any real sequence number.

closes odoo/enterprise#33256

Task: 2820482
Pr-enterprise: https://github.com/odoo/enterprise/pull/28326
X-original-commit: 124a56f0dce428e37430d70b945bb8c2c60a3641
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti (bib) <bib@odoo.com>