[FIX] sign: add default name to demo data
Since commit 860ace5013d7f5cde1018f5eb1ec4fce1f6b3537, if a sign
item does not have a name then the placeholder is not the type
placeholder anymore but false instead. This create an issue in
the demo data where the name was not previously set.

opw-2877149

closes odoo/enterprise#28303

Related: odoo/upgrade#3767
Signed-off-by: Arnaud Joset <arj@odoo.com>