[FIX] documents, *: restrict visibility for My Profile > Documents
When viewing documents via the My Profile > Documents button or via the
Partner > Documents button, a filter is applied on the Documents kanban view.
The user can in fact delete that filter to potentially see more file to which
he shouldn't have access to.

Step to reproduce the issue:
1) Install Documents and HR Documents modules
2) Once logged in go to My Profile > Documents and remove the filter
You can see file which you should not have access to via this view.

Solution: The solution is pretty simple. Instead of filtering the documents
via the Filter field we can simply filter the documents via the domain field.
This prevents the user from removing the field (as there are none).

*: documents_hr

opw-2858140

closes odoo/enterprise#29852

X-original-commit: d7cd0ec804749084f8885fcb5bd1ea94bf69ae63
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>
Signed-off-by: William Braeckman (wbr) <wbr@odoo.com>