[FIX] quality: alerts created with team company
How to reproduce the bug ?

- install quality
- create a second company
- in quality, create 2 quality teams, one for each company, with
different email aliases
- send an email to the address of the quality team of the created
company

What is the bug ?

When one needs to create a quality alert from an email, there is an
error if the quality team belongs to a company different from the "main"
one. This is due to the fact that when we create an alert, we are trying
to access its team.

opw-2781475

closes odoo/enterprise#28704

X-original-commit: 159309e7c17e0ef8274d89d90bcab1b97b40023b
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>