[FIX] l10n_be_reports: 281.50 commercial_partner should always be taken into account
The aim of this commit is to allow commercial_partner with a parent_id
to get a form 281.50.

Context:
A company partner is supposed to be completly distinct from its parent and thus
should have its own form 281.50.

Before this commit:
If a partner has a company and thus is a commercial_partner and has a parent,
it wouldn't get a 281.50 form created for it

After this commit:
The partner that is a company gets a form created for it even if it has a parent.

closes odoo/enterprise#29002

Task: 2820482
Enterprise-pr: https://github.com/odoo/enterprise/pull/28394
X-original-commit: a4f05355d14ecb9d5c6eea4dd28c5b0749d50746
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>