[FIX] planning: fix the traceback on setting the start_time on planning template
Issue1: Before this commit, when we enter the 24 as start_time then it
will generate a ValueError because the hour must be in 0..23.
There is a _sql_constraints on it but it accepts the 24 as time
while time max value should be 23:59.

so in thie commit, add the python constraint to restrict the 24 value
in start_time.

Issue2: Before this commit, when we enter the large integer value(100000000)
in shift template duration then it will generate error while
adding to current date as `date value out of range`

So in this commit fix this issue by adding the constraint on
duration field.

task-2873238

closes odoo/enterprise#29799

X-original-commit: 7202ff6aa4743c58c27f3869f38ec8a7beb5b13a
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>