[FIX] helpdesk_sale_loyalty: fix the helpdesk loyalty issue
Before this commit, the dropdown of coupon program does not
work properly and coupon program field also show promotions
in helpdesk ticket.

So in this commit, change the domain program_type as `coupons`
instead of 'coupon_program' as key is changed in refacoring and
also apply the default coupon type only in coupon wizard.

task-2858529

closes odoo/enterprise#33627

X-original-commit: 24275699c3daa7fb820fabee524e7ba501e12d8d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>