[REF] voip: always ensure a phonecall activity type exists
Currently, when calling ``create_call_in_queue`` an heuristic to find the
right activity type is performed. When no phonecall activity type is found
a fallback on Todo or a generic activity type is done. However when creating
activities, no ``voip.phonecall`` records are created when creating activities
under a type that is not phonecall. This leads to an UserError being raised
in ``create_call_in_queue``.

To avoid that issue, we now dynamically create the ``call`` activity type
if it has been removed or wrongly updated. That way voip always works as
intended instead of crashing or simply not working.

Task-2884301 (Voip synchronization simplification)

Part-of: odoo/enterprise#28421