[IMP] stock_barcode: better highlight
When a product is scanned, the quantity on the corresponding line is
updated, but for the user, the change is barely visible.
Now, the quantity will be write in bold if it still need to be processed
to catch the user's attention.
Same thing when a location is scanned (it will brievely flash).

task-2666478

Part-of: odoo/enterprise#27547