[FIX] sale_subscription: update upsell test value
Bedore this commit, the test would fail with the following message:
```
FAIL: TestSubscription.test_upsell_date_check
Traceback (most recent call last):
  File "/data/build/enterprise/sale_subscription/tests/test_sale_subscription.py", line 1347, in test_upsell_date_check
    self.assertEqual(upsell_invoice.amount_untaxed, 267.1, "The upsell amount should be equal to 267.1")
AssertionError: 252.05 != 267.1 : The upsell amount should be equal to 267.1
```

This bug was occuring because the discount was computed at two different dates.
During the second compute, "today" was another day and therefore the
time_to_invoice changed and therefore the real discount.

We decide to no recompute the discount if it has already a value.

closes odoo/enterprise#28509

Taskid: 2855699
X-original-commit: 8f22ab1b3886732c8ad117983765c6bcbd467a5f
Related: odoo/odoo#93859
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Signed-off-by: Arnaud Joset <arj@odoo.com>