[FIX] web_studio: properly display custom background on menu
When uploading a custom background image for the menu, the image is duplicated
to fit the screen size.

Step to reproduce the issue:
1) Install Studio
2) Go to the Home and go activate Studio
3) Go to Customizations > Change Background > Upload an image
You will see that the image is duplicated into the background

Solution: During the OWL migration [1], the class for custom background which
prevents the background to be repeated and to cover the entire screen has been
forgotten.

[1]: https://github.com/odoo/enterprise/commit/6563c16c2b2786efdbf175ae8254e242c10a9be7

opw-2819344

closes odoo/enterprise#33177

X-original-commit: b0bf230a60a4bd5a1550ae06b356476b547745db
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>