[FIX] web_studio: properly display custom background on menu
When uploading a custom background image for the menu, the image is duplicated
to fit the screen size.

Step to reproduce the issue:
1) Install Studio
2) Go to the Home and go activate Studio
3) Go to Customizations > Change Background > Upload an image
You will see that the image is duplicated into the background

Solution: During the OWL migration [1], the class for custom background which
prevents the background to be repeated and to cover the entire screen has been
forgotten.

[1]: https://github.com/odoo/enterprise/commit/6563c16c2b2786efdbf175ae8254e242c10a9be7

opw-2819344

closes odoo/enterprise#33221

X-original-commit: 7afaab08f0c5f8ff62f9c50d9663be066d8035c7
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>