[FIX] account_asset: depreciation schedule report
The Depreciation Schedule Report was not well displayed
in case of negative asset.

Let's assume a normal asset for 5 months, linear method,
500$ value and date today - 2 months (04-16-22), confirm
and validate.
-> 5 depreciation moves, the 2 first moves being posted.

The Report looks like this
| 01/01/2022 |     +    |    -   | 12/32/2022 | 01/01/2022 |     +    |    -   | 12/31/2022 |            |
|------------|----------|--------|------------|------------|----------|--------|------------|------------|
|   $ 0.00   | $ 500.00 | $ 0.00 |  $ 500.00  |   $ 0.00   | $ 200.00 | $ 0.00 |  $ 200.00  |  $ 300.00  |

Now, let's assume a negative asset in the same conditions
(method, duration, acquisition date), confirm and validate.

Go to the Report -> the line for negative asset is displayed the same
way as the positive one, it should be displayed this way
| 01/01/2022 |     +   |     -    | 12/32/2022 | 01/01/2022 |    +   |     -    | 12/31/2022 |           |
|------------|---------|----------|------------|------------|--------|----------|------------|-----------|
|   $ 0.00   |  $ 0.00 | $ 500.00 | $ -500.00  |   $ 0.00   | $ 0.00 | $ 200.00 | $ -200.00  | $ -300.00 |

The display when asset is closed is wrong too.

With this commit, we check if the asset comes from a refund,
if so, we recompute columns in the rigth way.

opw-2848330

closes odoo/enterprise#30960

X-original-commit: b7cab2f9c1e10aaf048cd48db162e6c60015394c
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>