[FIX] web_studio: allow to create a map view via Studio
Currently, you cannot create a map view for any module (e.g.: fleet). You get
a stack-trace instead.

Step to reproduce the issue:
1) Install the fleet and studio module
2) Go to Fleet then activate Studio
3) Go to Views > Map > Activate View
You get a stack trace.

Solution: During the migration to the OWL framework [1], a small mistake was
done inside the dialog when creating a new map view. The `bodyTemplate` was
wrongly instanciated. Furthermore, the default value for the field has been
set to mimic the behavior in 14.0.

[1]: https://github.com/odoo/enterprise/commit/6563c16c2b2786efdbf175ae8254e242c10a9be7#diff-3770e56f0cadfcf93d145fc410c499aa257d19ac8d12655ba35407b117eb757b

opw-2881912

closes odoo/enterprise#29503

X-original-commit: 2781321c82f1ddc8e60c08e761df3712f37373b5
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>