[IMP] stock_barcode: WOWL form view
Before this commit, when a line was edited, the displayed form view was
shown thank to the `view_widget` and the `ViewsWidgetAdapter` component.
But now, a form view can easily be displayed in a component thank to the
`View` component.
That means some adjustments are done here, like converting `digipad` and
`set_reserved_qty_button` widgets into OWL's component.

task-2885757

Part-of: odoo/enterprise#30045