[FIX] sign: no false placeholder displayed in sign items
Steps to reproduce:

- Install sign
- Create a template and add a sign item
- Set an empty placeholder with the developer mode
on the sign item and save

Expected behavior:
The placeholder is empty

Current behavior:
The placeholder is false

Explanation:
It was thought in a previous fix 860ace5013d7f5cde1018f5eb1ec4fce1f6b3537
that the false value was the expected behavior. But is was just inudced by
an empty placeholder in the sign item type that returned the false value.
To solve the issue we add || "" after the different places where the name
is set so that if the type placeholder is false then the sign item
placeholder is an empty string.

opw-2767417

Part-of: odoo/enterprise#28303