[FIX] l10n_co_edi: fixed taxes correctly handled
How to reproduce the bug ?

- install account_accountant, l10n_co, l10n_co_edi
- follow the procedure to test l10n_co
(https://docs.google.com/document/d/167i3L1_AKV-6VS8yYevYaObja24Sx5NZ8klTg7eUWjg)
- Go to Accounting > Configuration > Taxes and click on the one called
"IVA Ventas 19%"
- Change the field "Tax Computation" to "Fixed", (in "Advanced Options")
"Tax Group" to "IC" and "Tipo de Valor" to "IC". Save the modification.
- Go to Customers > Invoices and create a new invoice.
- Pick the customer you have modified in the test procedure and add a
product to the invoice.
- Save and Confirm the invoice.

What is the bug ?

With a colombian company, when you create an invoice with a product
with a fixed value tax you will get an error on confirmation. This is
due to the fact that the method computing the taxes
(_l10n_co_edi_get_round_amount) doesn't work with the value specified
with a fixed value.

opw-2883894

closes odoo/enterprise#28616

X-original-commit: 558bf576abbb6ab8563c755b4a3549e3a51be502
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>