[FIX] sign: sign item placeholder is type default if empty
Steps to reproduce:
- Create a sign item
- Set an empty placeholder with developer mode
- Refresh
- Click on the sign item you just created

Expected behavior:
The placeholder is empty

Current behavior:
The placeholder is the sign item type default placeholder

Explanation:
The javascript is unabled to make the difference between the
name of a sign item that was just created (and has a default
value of false) and a sign item that had its placeholder set
to false. We solve that in the backend by setting a default
value for the sign item name at creation. It's cleaner than
the current implementation as the default placeholder is only
set once and it makes more sense.

opw-2767417

Part-of: odoo/enterprise#28303