[REF] sale: simplify price computation logic
Adapt to community changes in sale simplifying and clarifying the price
computation code & logic.

The custom logic of price_computing_kwargs is now moved to sale_temporal
because it's not necessary in community and wouldn't integrate well in the
standard sale behavior now that it's been improved to reduce operations as much
as possible.

closes odoo/enterprise#28552

Related: odoo/odoo#88230
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>