[FIX] account_intrastat: enable product weight/volume
The product's weight is utilised on the intrastat report, but the
section used to define these values on the product is actually only
accesible when the stock module is installed. To fix this, this commit
adds the user group 'account.group_account_invoice' to the inventory
page on the product page.

closes odoo/enterprise#29136

Task-id: 2884474
X-original-commit: 0b0fdec73908941fdda3ae7eea5fe2ae5a1dd562
Signed-off-by: Brice Bartoletti <bib@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>