[IMP] account_intrastat: consolidate system/type
On the intrastat report the system and type fields represent the same
thing and are thus redundant. It therefore makes sense to consolidate
them into one field called 'Commodity Flow'.

In this commit the two fields are instead represented as one field, with
the 'system' first, and the 'type' following it in brackets e.g.
19 (Dispatch).

The xlsx export should instead contain only the 'system' (the 'type' is
actually redundant when considering the 'system' codes), so the
'print_xlsx' function is modified such that only the 'system' is used.
In order to achieve this distinction, a key 'country_format' is
inserted into the options inside of the 'print_xlsx'. The content of
this report 'Commodity Flow' field is generated depending on this key.

The tests are updated to match this new structure and a new test has
been added to test the content of the report under the same conditions
as an xlsx export (i.e. codes instead of names).

closes odoo/enterprise#29506

Task-id: 2884333
Signed-off-by: Brice Bartoletti <bib@odoo.com>