[FIX] l10n_mx_edi: _l10n_mx_edi_get_tax_objected only existed in the 4.0 module
With the merge for cfdi 4.0 in https://github.com/odoo/enterprise/pull/27675,
we needed the tax_objected value already in l10n_mx_edi, while it was
only defined in l10n_mx_edi_40.  This broke the sending with only the old
modules installed. (not the cfdi 4.0 ones)

That is why we return a default value 02 for the method in l10n_mx_edi
which means tax breakdown should be done. (which is always the case for cfdi v3.3 anyways)

opw-2888693

closes odoo/enterprise#28855

X-original-commit: 6e8048a63ebe20d093c5f63e06dab0248ff79fbf
Signed-off-by: William André (wan) <wan@odoo.com>