[IMP] stock_barcode{_picking_batch}: add user in kanban
Currently in the kanban view in barcode, there is no way to know if a picking is already started.
As a result, if several users use the app, there is a risk of having 2 people doing the same thing.
Therefore we add in the responsible user into the kanban view and auto-filter on revelant pickings
(i.e. ones assigned to current user or not assigned to anyone) to avoid this issue.
Additionally we reassign the responsible user for pickings to the current user
if it appears a picking is being taking over by another user.

TaskId: 2871679
COM PR: odoo/odoo#94633

closes odoo/enterprise#31740

X-original-commit: 5e76c81deedfce68f98e0c5030387e1ede8895e8
Related: odoo/odoo#101026
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: Ahmed Khalaf  <ahkh@odoo.com>