[FIX] l10n_mx_edi: forgot to add method for the old templates
In PR https://github.com/odoo/enterprise/pull/28578 we did
a little refactoring to support the new cfdi 40 and forgot
to provide templates for the old methods as _l10n_mx_edi_get_invoice_templates
was only implemented in l10n_mx_edi_40 and not in l10n_mx_edi.

This avoids a traceback in v15.0 when not installing the new modules
and sending invoices.  (as it would not find back the method)

closes odoo/enterprise#28692

X-original-commit: dae5404d15059575e295c36b20f8544b19311da7
Signed-off-by: William André (wan) <wan@odoo.com>