[IMP] payment_sepa_direct_debit: disable tokens of disabled acquirers
Before this commit, zombie payment tokens (= tokens linked to disabled
acquirers) could still
1) be used by internal users and
2) reactivated when the acquirer’s state changed to ‘test’ or ‘enabled’.
This is not desirable because zombie tokens should neither be used,
nor reactivated.

After this commit, all tokens related to an acquirer are unassigned from
linked documents and archived as soon as the acquirer’s state is changed
to ‘disabled’. Creating a payment with an archived token is prohibited.
In addition, archived tokens cannot be un-archived anymore.

task-2649806

closes odoo/enterprise#28661

Related: odoo/odoo#93774
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>