[FIX] hr_expense_extract: auto OCR works for expense created by mail
How to reproduce the bug ?

- Install hr_exepense
- If needed, change the email address of the current user to the one
that will send the email.
- In Settings > Expenses, make sure that an alias (check that incoming
emails create expenses) is set and that Digitalize automatically is
selected.
- Send an email to the address set earlier.
- Wait for the cron to fetch emails or execute the cron manually.
- In the Expenses app, click on on the expense that has the subject of
the email as description

What is the bug ?

When you create an expense by sending an email with an attachment, the
expense will be created but the attachment will not be automatically
digitalized even if the option is set in the settings.
This PR partially revert the following commit https://github.com/odoo/enterprise/commit/dded12fa892cb4a6f764494adff140a422e522b9

opw-2869381

closes odoo/enterprise#28826

X-original-commit: d84e9afddc730818dad7d646c0c0f70a2ed29007
Signed-off-by: Adrien Minet <admi@odoo.com>
Signed-off-by: Kevin Baptiste <kba@odoo.com>
Signed-off-by: Minet Adrien (admi) <admi@odoo.com>