[FIX] account_intrastat: remove date, use full country names
For the readability of the intrastat report, the date (which is not
necessary) is removed, and the countries are displayed using their
country names instead of their 2-digit ISO codes. The 2-digit ISO codes
are still used in the declarations and the xlsx export.

Tests are modified in account_intrastat and account_intrastat_expiry to
comply to the new format.

closes odoo/enterprise#29285

Task-id: 2884387
X-original-commit: 0422afc0997dfcb8ab3bf3c4d76ea520a20b8b90
Signed-off-by: Brice Bartoletti <bib@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>