[FIX] l10n_nl_intrastat: remove region code field
The region code field on the report is useless in the context of the
Netherlands (it will always be empty).

This commit changes the function by which the columns are included by
introducing a function '_show_region_code()' in account_intrastat, which
is overrided in l10n_nl_intrastat. The function is employed inside of
the _create_intrastat_report_line function and the _get_columns_names
functions, in order to decide whether to include the region code
column/values when generating the report.

task_id: 2884381

closes odoo/enterprise#29189

X-original-commit: 42e23a0ec7602c9e089c4dc0fb3acecb5bcaabbd
Signed-off-by: Brice Bartoletti <bib@odoo.com>
Signed-off-by: Daniel Kosky (dako) <dako@odoo.com>