[IMP] mail_enterprise, *: replace createChatterContainer during tests
*: approvals, voip.

The createChatterContainer helper was used during tests to mount
a chatter container and test it. However, this approach is not
very realistic and blocks some waiting PRs. In order to get closer
from the reality, let's mount a webClient and open a form view
containing a chatter.

closes odoo/enterprise#28684

Related: odoo/odoo#94175
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>